+++ src/SMAPI/Framework/Models/SConfig.cs
@@ -15,20 +15,21 @@ internal class SConfig
     /// <summary>The default config values, for fields that should be logged if different.</summary>
     private static readonly IDictionary<string, object> DefaultValues = new Dictionary<string, object>
     {
-        [nameof(CheckForUpdates)] = true,
-        [nameof(CheckForBlacklistUpdates)] = true,
-        [nameof(CheckContentIntegrity)] = true,
-        [nameof(ListenForConsoleInput)] = true,
-        [nameof(ParanoidWarnings)] = Constants.IsDebugBuild,
-        [nameof(UseBetaChannel)] = Constants.ApiVersion.IsPrerelease(),
-        [nameof(GitHubProjectName)] = "Pathoschild/SMAPI",
-        [nameof(WebApiBaseUrl)] = "https://smapi.io/api/",
-        [nameof(LogNetworkTraffic)] = false,
-        [nameof(LogTechnicalDetailsForBrokenMods)] = false,
-        [nameof(RewriteMods)] = true,
-        [nameof(FixHarmony)] = true,
-        [nameof(UseCaseInsensitivePaths)] = Constants.Platform is Platform.Android or Platform.Linux,
-        [nameof(SuppressHarmonyDebugMode)] = true
+        [nameof(SConfig.CheckForUpdates)] = true,
+        [nameof(SConfig.CheckForBlacklistUpdates)] = true,
+        [nameof(SConfig.CheckContentIntegrity)] = true,
+        [nameof(SConfig.ListenForConsoleInput)] = true,
+        [nameof(SConfig.ParanoidWarnings)] = Constants.IsDebugBuild,
+        [nameof(SConfig.UseBetaChannel)] = Constants.ApiVersion.IsPrerelease(),
+        [nameof(SConfig.GitHubProjectName)] = "Pathoschild/SMAPI",
+        [nameof(SConfig.WebApiBaseUrl)] = "https://smapi.io/api/",
+        [nameof(SConfig.LogNetworkTraffic)] = false,
+        [nameof(SConfig.LogTechnicalDetailsForBrokenMods)] = false,
+        [nameof(SConfig.RewriteMods)] = true,
+        [nameof(SConfig.EnableConfigMenu)] = true,
+        [nameof(SConfig.FixHarmony)] = true,
+        [nameof(SConfig.UseCaseInsensitivePaths)] = Constants.Platform is Platform.Android or Platform.Linux,
+        [nameof(SConfig.SuppressHarmonyDebugMode)] = true
     };
 
     /// <summary>The default values for <see cref="SuppressUpdateChecks"/>, to log changes if different.</summary>
@@ -46,7 +47,7 @@ internal class SConfig
     // Note: properties must be writable to support merging config.user.json into it.
     //
 
-    /// <summary>Whether to enable development features.</summary>
+    /// <summary>Whether to show much more info in the SMAPI consoler window, intended for mod developers. Not recommended for most players.</summary>
     public bool DeveloperMode { get; set; }
 
     /// <summary>Whether to check for newer versions of SMAPI and mods on startup.</summary>
@@ -61,7 +62,7 @@ internal class SConfig
     /// <summary>Whether SMAPI should listen for console input to support console commands.</summary>
     public bool ListenForConsoleInput { get; set; }
 
-    /// <summary>Whether to add a section to the 'mod issues' list for mods which which directly use potentially sensitive .NET APIs like file or shell access.</summary>
+    /// <summary>Whether to add a section to the 'mod issues' list for mods which directly use potentially sensitive .NET APIs like file or shell access.</summary>
     public bool ParanoidWarnings { get; set; }
 
     /// <summary>Whether to show beta versions as valid updates.</summary>
@@ -82,6 +83,9 @@ internal class SConfig
 
     /// <summary>Whether SMAPI should rewrite mods for compatibility.</summary>
     public bool RewriteMods { get; set; }
+
+    /// <summary>Whether SMAPI should add an in-game config menu through Generic Mod Config Menu.</summary>
+    public bool EnableConfigMenu { get; set; }
 
     /// <summary>Whether to apply fixes to Harmony so it works with Stardew Valley.</summary>
     public bool FixHarmony { get; set; }
@@ -130,6 +134,7 @@ internal class SConfig
     /// <param name="blacklistUrl"><inheritdoc cref="BlacklistUrl" path="/summary" /></param>
     /// <param name="verboseLogging"><inheritdoc cref="VerboseLogging" path="/summary" /></param>
     /// <param name="rewriteMods"><inheritdoc cref="RewriteMods" path="/summary" /></param>
+    /// <param name="enableConfigMenu"><inheritdoc cref="EnableConfigMenu" path="/summary" /></param>
     /// <param name="fixHarmony"><inheritdoc cref="FixHarmony" path="/summary" /></param>
     /// <param name="useCaseInsensitivePaths"><inheritdoc cref="UseCaseInsensitivePaths" path="/summary" /></param>
     /// <param name="logNetworkTraffic"><inheritdoc cref="LogNetworkTraffic" path="/summary" /></param>
@@ -153,6 +158,7 @@ internal class SConfig
         string blacklistUrl,
         string[]? verboseLogging,
         bool? rewriteMods,
+        bool? enableConfigMenu,
         bool? fixHarmony,
         bool? useCaseInsensitivePaths,
         bool? logNetworkTraffic,
@@ -177,6 +183,7 @@ internal class SConfig
         this.BlacklistUrl = blacklistUrl;
         this.VerboseLogging = new HashSet<string>(verboseLogging ?? [], StringComparer.OrdinalIgnoreCase);
         this.RewriteMods = rewriteMods ?? (bool)SConfig.DefaultValues[nameof(this.RewriteMods)];
+        this.EnableConfigMenu = enableConfigMenu ?? (bool)SConfig.DefaultValues[nameof(this.EnableConfigMenu)];
         this.FixHarmony = fixHarmony ?? (bool)SConfig.DefaultValues[nameof(this.FixHarmony)];
         this.UseCaseInsensitivePaths = useCaseInsensitivePaths ?? (bool)SConfig.DefaultValues[nameof(this.UseCaseInsensitivePaths)];
         this.LogNetworkTraffic = logNetworkTraffic ?? (bool)SConfig.DefaultValues[nameof(this.LogNetworkTraffic)];